python-dash/3-devcontainers/requirements.txt
dash
pandas
dash_bootstrap_components
psycopg2-binary
python-dotenv
sqlalchemy